# docs/swagger.yaml
definitions:
  channel.Channel:
    properties:
      created_at:
        description: CreatedAt 创建时间
        type: integer
      id:
        description: Id 频道ID
        type: integer
      name:
        description: Name 频道名称
        type: string
      secret:
        description: Secret 频道密钥
        type: string
      token:
        description: Token 频道token
        type: string
      updated_at:
        description: UpdatedAt 更新时间
        type: integer
      url:
        description: Url 频道地址
        type: string
      user_id:
        description: UserId 用户ID
        type: integer
    type: object
  channel.Channels:
    properties:
      items:
        items:
          $ref: '#/definitions/channel.Channel'
        type: array
      total:
        type: integer
    type: object
  channel.CreateReq:
    properties:
      name:
        type: string
      secret:
        type: string
      url:
        type: string
    required:
    - name
    type: object
  channel.QueryReq:
    properties:
      keywords:
        type: string
      page_number:
        type: integer
      page_size:
        type: integer
    type: object
  channel.UpdateReq:
    properties:
      name:
        type: string
      secret:
        type: string
      url:
        type: string
    type: object
  message.ProducerReq:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  token.LoginReq:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
  token.Tk:
    properties:
      access_token:
        description: 登录token
        type: string
      refresh_token:
        description: 刷新token
        type: string
    required:
    - access_token
    - refresh_token
    type: object
  user.CreateReq:
    properties:
      password:
        description: 密码
        maxLength: 20
        minLength: 6
        type: string
      username:
        description: 用户名
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.Role:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - RoleAnonymous
    - RoleAdmin
  user.Status:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - StatusNormal
    - StatusDeleted
  user.User:
    properties:
      created_at:
        description: CreatedAt 创建时间
        type: integer
      deleted_at:
        description: DeletedAt 删除时间
        type: integer
      email:
        description: Email 邮箱
        type: string
      id:
        description: ID 用户ID
        type: integer
      nickname:
        description: Nickname 昵称
        type: string
      password:
        description: Password 密码
        type: string
      phone:
        description: Phone 手机号
        type: string
      role:
        allOf:
        - $ref: '#/definitions/user.Role'
        description: Role 角色
      status:
        allOf:
        - $ref: '#/definitions/user.Status'
        description: Status 状态
      updated_at:
        description: UpdatedAt 更新时间
        type: integer
      username:
        description: Username 用户名
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/channel:
    post:
      consumes:
      - application/json
      description: 创建频道
      parameters:
      - description: 创建频道请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/channel.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.Channel'
      summary: 创建频道
      tags:
      - channel
  /api/v1/channel/{id}:
    post:
      consumes:
      - application/json
      description: 查询频道
      parameters:
      - description: 查询频道请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/channel.QueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.Channels'
      summary: 查询频道
      tags:
      - channel
    put:
      consumes:
      - application/json
      description: 更新频道
      parameters:
      - description: 更新频道请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/channel.UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 更新频道
      tags:
      - channel
  /api/v1/channel/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除频道
      parameters:
      - description: 频道ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 删除频道
      tags:
      - channel
  /api/v1/message:
    post:
      consumes:
      - application/json
      description: 客户端发送消息
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/message.ProducerReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
      summary: 客户端发送消息
      tags:
      - message
  /api/v1/token/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/token.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/token.Tk'
      summary: 登录
      tags:
      - token
  /api/v1/token/logout:
    post:
      consumes:
      - application/json
      description: 登出
      parameters:
      - description: 登出请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/token.Tk'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 登出
      tags:
      - token
  /api/v1/token/refresh:
    post:
      consumes:
      - application/json
      description: 刷新
      parameters:
      - description: 刷新请求参数
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/token.Tk'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 刷新
      tags:
      - token
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 创建用户请求参数
        in: body
        name: Object
        schema:
          $ref: '#/definitions/user.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      summary: 创建用户
      tags:
      - 用户
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 删除用户
      tags:
      - 用户
    get:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      summary: 查询用户
      tags:
      - 用户
swagger: "2.0"